+++ src/main/java/com/github/misosouptgit/instantnbt/mixin/CompoundTagMixin.java
@@ -1,0 +1,25 @@
+package com.github.misosouptgit.instantnbt.mixin;
+
+import com.github.misosouptgit.instantnbt.ownership.TagWriteHooks;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.Tag;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+/**
+ * Marks tracked OwnedTags dirty on compound mutation (Project Plan: minimal Mixin surface).
+ */
+@Mixin(CompoundTag.class)
+public abstract class CompoundTagMixin {
+	@Inject(method = "put(Ljava/lang/String;Lnet/minecraft/nbt/Tag;)Lnet/minecraft/nbt/Tag;", at = @At("HEAD"))
+	private void instantnbt$put(String key, Tag value, CallbackInfoReturnable<Tag> cir) {
+		TagWriteHooks.onMutate((Tag) (Object) this);
+	}
+
+	@Inject(method = "remove(Ljava/lang/String;)V", at = @At("HEAD"))
+	private void instantnbt$remove(String key, org.spongepowered.asm.mixin.injection.callback.CallbackInfo ci) {
+		TagWriteHooks.onMutate((Tag) (Object) this);
+	}
+}